packages/ai-core/pyproject.toml
[project]
name = "ai-core"
version = "0.1.0"
description = "Vendor-agnostic LLM client abstractions for AI Stack"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "AI Stack Team" }
]

dependencies = [
    # LLM Providers
    "openai>=1.58.1",
    "anthropic>=0.42.0",
    "google-generativeai>=0.8.3",
    "ollama>=0.4.4",
    
    # Utilities
    "httpx>=0.28.1",
    "tiktoken>=0.8.0",
    "pydantic>=2.12.5",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
